create close description command
